Rebased to master, fixed some things that broke as a result.
There were some build problems after rebasing. Some were due to missing
types in the test file, but main problem was that the jest-fetch-mock
doesn't have a types file and the type defined for fetch in eosjs2-jsonrpc
of course doesn't define the extra mocking related methods that the
mock fetch has available.

Solution was to simply not build the .test.ts files as part of the
distribution build, which is the right thing to do anyway.