use native textEncoder/Decoder
'text-encoding' npm package is no longer maintained and the node native implementation is available since 8.3
https://nodejs.org/api/util.html#util_class_util_textencoder
https://nodejs.org/api/util.html#util_class_util_textdecoder

I would suggest that this package should automatically import and use these native functions when it detects that it's being run in node. Manual import should only be necessary if you need to override this.